% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP.R
\name{AHP}
\alias{AHP}
\title{AHP: Analytic Hierarchy Process}
\usage{
AHP(data)
}
\arguments{
\item{data}{a numeric matrix}
}
\value{
a list
}
\description{
AHP is a multi-objective decision analysis method that can be used to determine indicator weights.
}
\examples{
data<-matrix(c(1,1/2,1/6,2,1,1/4,6,4,1),3,3)
AHP(data)
}
