% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_CV.R
\name{cluster_CV}
\alias{cluster_CV}
\title{cluster_CV:R-type system clustering combined with coefficient of variation method}
\usage{
cluster_CV(data, k)
}
\arguments{
\item{data}{a numeric data frame or tibble}

\item{k}{a number}
}
\value{
a list
}
\description{
cluster_CV is a methods of screening evaluation indicators.
}
\examples{
data<-as.data.frame(t(iris[,-5]))
k<-2
cluster_CV(data,k)
}
