% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{EM:Entropy Method}
\usage{
EM(data, type)
}
\arguments{
\item{data}{a numeric data frame or tibble}

\item{type}{a numeric vector}
}
\value{
a data frame
}
\description{
EM is a method that can be used to determine indicator weights.
}
\examples{
data<-iris[,-5]
type<-c(1,2,1,1)
EM(data,type)
}
