% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuzzy.R
\name{Fuzzy}
\alias{Fuzzy}
\title{Fuzzy:Fuzzy Comprehension Evaluation Method}
\usage{
Fuzzy(r, w, v, s)
}
\arguments{
\item{r}{a numeric matrix}

\item{w}{a numeric vector}

\item{v}{a vector}

\item{s}{a numeric vector}
}
\value{
a list
}
\description{
Fuzzy is a analysis method based on fuzzy mathematics for comprehensive evaluation.
}
\examples{
r<-matrix(c(0.1029,0.3088,0.4853,0.1029,
            0.1764,0.2941,0.4216,0.1078,
            0.1471,0.3382,0.4118,0.1029),3,4,
            dimnames=list(c("Z1","Z2","Z3"),c("A","B","C","D")))
w<-c(0.1,0.3,0.6)
v<-c("A","B","C","D")
s<-c(100,80,60,40)
Fuzzy(r,w,v,s)
}
