% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRA.R
\name{GRA}
\alias{GRA}
\title{GRA: Grey Relational Analysis}
\usage{
GRA(data, r, w)
}
\arguments{
\item{data}{a numeric matrix}

\item{r}{a number}

\item{w}{a numeric vector}
}
\value{
a list
}
\description{
GRA is a multivariate statistical analysis method for comprehensive evaluation.
}
\examples{
r=0.4
w=c(0.2,0.2,0.3,0.1,0.1,0.1)
data<-matrix(c(1988,2061,2335,2750,3356,3806,
               386,408,422,482,511,561,
               839,846,960,1258,1577,1893,
               763,808,953,1010,1268,1352),
               4,6,byrow = TRUE,
               dimnames = list(c("A0","A1","A2","A3"),
                               c("Z1","Z2","Z3","Z4","Z5","Z6")))
GRA(data,r,w)
}
