% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\title{norm:Normalization}
\usage{
norm(data, type)
}
\arguments{
\item{data}{a numeric data frame or tibble or matrix}

\item{type}{a numeric vector}
}
\value{
a normalized data frame
}
\description{
Norm is a method for normalizing positive and negative indicators.
}
\examples{
data<-iris[1:10,-5]
type<-c(1,1,1,1)
norm(data,type)
}
