% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA:Principal Component Analysis}
\usage{
PCA(data)
}
\arguments{
\item{data}{a numeric data frame or tibble}
}
\value{
a list
}
\description{
PCA is a statistical method of dimension reduction that can be used to determine indicator weights.
}
\examples{
data<-iris[,-5]
PCA(data)
}
