% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPSIS.R
\name{TOPSIS}
\alias{TOPSIS}
\title{TOPSIS:Technique for Order Preference by Similarity to an Ideal Solution}
\usage{
TOPSIS(data, w, type)
}
\arguments{
\item{data}{a numeric data frame or tibble}

\item{w}{a numeric vector}

\item{type}{a numeric vector}
}
\value{
a data frame
}
\description{
TOPSIS is a ranking method approximating to the ideal solution for comprehensive evaluation.
}
\examples{
data<-iris[,-5]
w<-c(0.2,0.1,0.3,0.4)
type<-c(1,1,1,1)
TOPSIS(data,w,type)
}
